+++ src/test/java/LoginTest.java
@@ -1,0 +1,109 @@
+import io.github.bonigarcia.wdm.WebDriverManager;
+import io.qameta.allure.junit4.DisplayName;
+import org.example.PageObject.LoginPage;
+import org.example.PageObject.MainPage;
+import org.example.PageObject.RegisterPage;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.FixMethodOrder;
+import org.junit.Test;
+import org.junit.runners.MethodSorters;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+
+import java.time.Duration;
+
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
+public class LoginTest {
+    private MainPage objMainPage;
+    private LoginPage objLoginPage;
+    private RegisterPage objRegisterPage;
+    private WebDriver driver;
+
+    @Before
+    public void startUp() {
+        WebDriverManager.chromedriver().setup();//драйвер хром
+        driver = new ChromeDriver();//экземпляр драйвера хром
+
+        //замена на драйвер яндекс.браузера (закомментить экземпляр драйвера хрома выше и раскомментить строчки снизу)
+//        ChromeOptions options = new ChromeOptions();
+//        System.setProperty("webdriver.chrome.driver", "C:\\Users\\DNS\\yandexdriver.exe");
+//        options.setBinary("C:\\Users\\DNS\\AppData\\Local\\Yandex\\YandexBrowser\\Application\\browser.exe");
+//        driver = new ChromeDriver(options);
+
+        objMainPage = new MainPage(driver);//новый объект класса главной страницы
+        objLoginPage = new LoginPage(driver);//новый объект класса страницы "Вход"
+        objRegisterPage = new RegisterPage(driver);//новый объект класса страницы "Регистрация"
+        objMainPage.open();//открытие тестовой страницы
+        objMainPage.waitForLoadPage();//ожидание загрузки
+    }
+
+    @Test
+    @DisplayName("Проверка входа по кнопке Войти в аккаунт на главной")
+    public void cSuccessfulLoginWithButtonEnterInAccount() {
+        objMainPage.clickLoginAccountButton();//клик на "Войти в аккаунт"
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));//ожидание открытия страницы "Вход"
+        objLoginPage.waitForLoadPage();//ожидание появления окна "Вход"
+        objLoginPage.isLoginPageWindowOpen();//проверка открытия окна входа
+        objLoginPage.inputEmail();//ввод почты
+        objLoginPage.inputPasswordWithSixChars();//ввод пароля
+        objLoginPage.clickEnterButton();//клик на "Войти"
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));//ожидание открытия главной страницы
+        objMainPage.waitForLoadPage();//ожидание появления окна главной страницы
+        objMainPage.isSendOrderButtonPresent();//проверка наличия кнопки "Оформить заказ"
+    }
+
+    @Test
+    @DisplayName("Проверка входа через кнопку Личный кабинет")
+    public void dSuccessfulLoginWithAccountButton() {
+        objMainPage.clickAccountButton();//клик на "Личный кабинет"
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));//ожидание открытия страницы "Вход"
+        objLoginPage.waitForLoadPage();//ожидание появления окна "Вход"
+        objLoginPage.isLoginPageWindowOpen();//проверка открытия окна входа
+        objLoginPage.inputEmail();//ввод почты
+        objLoginPage.inputPasswordWithSixChars();//ввод пароля
+        objLoginPage.clickEnterButton();//клик на "Войти"
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));//ожидание открытия главной страницы
+        objMainPage.waitForLoadPage();//ожидание появления окна главной страницы
+        objMainPage.isSendOrderButtonPresent();//проверка наличия кнопки "Оформить заказ"
+    }
+
+    @Test
+    @DisplayName("Проверка входа через кнопку в форме регистрации")
+    public void eSuccessfulLoginWithRegisterPageButton() {
+        objMainPage.clickLoginAccountButton();//клик на "Войти в аккаунт"
+        objLoginPage.clickRegisterButton();//клик на "Зарегистрироваться"
+        objRegisterPage.clickEnterButton();//клик на "Войти"
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));//ожидание открытия страницы "Вход"
+        objLoginPage.waitForLoadPage();//ожидание появления окна "Вход"
+        objLoginPage.isLoginPageWindowOpen();//проверка открытия окна входа
+        objLoginPage.inputEmail();//ввод почты
+        objLoginPage.inputPasswordWithSixChars();//ввод пароля
+        objLoginPage.clickEnterButton();//клик на "Войти"
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));//ожидание открытия главной страницы
+        objMainPage.waitForLoadPage();//ожидание появления окна главной страницы
+        objMainPage.isSendOrderButtonPresent();//проверка наличия кнопки "Оформить заказ"
+    }
+
+    @Test
+    @DisplayName("Проверка входа через кнопку в форме восстановления пароля")
+    public void fSuccessfulLoginWithResetPasswordPageButton() {
+        objMainPage.clickLoginAccountButton();//клик на "Войти в аккаунт"
+        objLoginPage.clickForgotButton();//клик на "Восстановить пароль"
+        objRegisterPage.clickEnterButton();//клик на "Войти"
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));//ожидание открытия страницы "Вход"
+        objLoginPage.waitForLoadPage();//ожидание появления окна "Вход"
+        objLoginPage.isLoginPageWindowOpen();//проверка открытия окна входа
+        objLoginPage.inputEmail();//ввод почты
+        objLoginPage.inputPasswordWithSixChars();//ввод пароля
+        objLoginPage.clickEnterButton();//клик на "Войти"
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));//ожидание открытия главной страницы
+        objMainPage.waitForLoadPage();//ожидание появления окна главной страницы
+        objMainPage.isSendOrderButtonPresent();//проверка наличия кнопки "Оформить заказ"
+    }
+
+    @After
+    public void cleanUp() {
+        driver.quit();
+    }
+}
