+++ src/test/java/RegistrationTest.java
@@ -1,0 +1,71 @@
+import io.github.bonigarcia.wdm.WebDriverManager;
+import io.qameta.allure.junit4.DisplayName;
+import org.example.PageObject.LoginPage;
+import org.example.PageObject.MainPage;
+import org.example.PageObject.RegisterPage;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.FixMethodOrder;
+import org.junit.Test;
+import org.junit.runners.MethodSorters;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+
+import java.time.Duration;
+
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
+public class RegistrationTest {
+    private MainPage objMainPage;
+    private LoginPage objLoginPage;
+    private RegisterPage objRegisterPage;
+    private WebDriver driver;
+
+    @Before
+    public void startUp() {
+        WebDriverManager.chromedriver().setup();//драйвер хром
+        driver = new ChromeDriver();//экземпляр драйвера хром
+
+        //замена на драйвер яндекс.браузера (закомментить экземпляр драйвера хрома выше и раскомментить строчки снизу)
+//        ChromeOptions options = new ChromeOptions();
+//        System.setProperty("webdriver.chrome.driver", "C:\\Users\\DNS\\yandexdriver.exe");
+//        options.setBinary("C:\\Users\\DNS\\AppData\\Local\\Yandex\\YandexBrowser\\Application\\browser.exe");
+//        driver = new ChromeDriver(options);
+
+        objMainPage = new MainPage(driver);//новый объект класса главной страницы
+        objLoginPage = new LoginPage(driver);//новый объект класса страницы "Вход"
+        objRegisterPage = new RegisterPage(driver);//новый объект класса страницы "Регистрация"
+        objMainPage.open();//открытие тестовой страницы
+        objMainPage.waitForLoadPage();//ожидание загрузки
+    }
+
+    @Test
+    @DisplayName("Проверка успешной регистрации пользователя")
+    public void aSuccessfulRegistration() {
+        objMainPage.clickLoginAccountButton();//клик на "Войти в аккаунт"
+        objLoginPage.clickRegisterButton();//клик на "Зарегистрироваться"
+        objRegisterPage.inputName();//ввод имени
+        objRegisterPage.inputEmail();//ввод почты
+        objRegisterPage.inputPasswordWithSixChars();//ввод пароля
+        objRegisterPage.clickRegisterButton();//клик на "Зарегистрироваться"
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));//ожидание открытия страницы "Вход"
+        objLoginPage.waitForLoadPage();//ожидание появления окна "Вход"
+        objLoginPage.isLoginPageWindowOpen();//проверка открытия окна входа
+    }
+
+    @Test
+    @DisplayName("Проверка ошибки для некорректного пароля меньше шести символов")
+    public void bCheckPasswordErrorIfLessSixChars() {
+        objMainPage.clickLoginAccountButton();//клик на "Войти в аккаунт"
+        objLoginPage.clickRegisterButton();//клик на "Зарегистрироваться"
+        objRegisterPage.inputName();//ввод имени
+        objRegisterPage.inputEmail();//ввод почты
+        objRegisterPage.inputPasswordWithFiveChars();//ввод пароля
+        objRegisterPage.clickRegisterButton();//клик на "Зарегистрироваться"
+        objRegisterPage.isPasswordErrorPresent();//проверка появления ошибки
+    }
+
+    @After
+    public void cleanUp() {
+        driver.quit();
+    }
+}
